fmt + split off tests from mod.rs to tests.rs
