+++ crates/chibi-core/src/state/entries.rs
@@ -4,8 +4,8 @@
 //! using the builder pattern.
 
 use crate::context::{
-    EntryMetadata, TranscriptEntry, ENTRY_TYPE_ARCHIVAL, ENTRY_TYPE_COMPACTION,
-    ENTRY_TYPE_CONTEXT_CREATED, ENTRY_TYPE_MESSAGE, ENTRY_TYPE_TOOL_CALL, ENTRY_TYPE_TOOL_RESULT,
+    ENTRY_TYPE_ARCHIVAL, ENTRY_TYPE_COMPACTION, ENTRY_TYPE_CONTEXT_CREATED, ENTRY_TYPE_MESSAGE,
+    ENTRY_TYPE_TOOL_CALL, ENTRY_TYPE_TOOL_RESULT, EntryMetadata, TranscriptEntry,
 };
 
 /// Create a transcript entry for a user message
